+++ docs/examples/03_backends_ros/files/01_ros_hello_world_talker.py
@@ -5,16 +5,12 @@ from roslibpy import Topic
 
 from compas_fab.backends import RosClient
 
-client = RosClient()
-client.run()
+with RosClient() as client:
+    talker = Topic(client, '/chatter', 'std_msgs/String')
 
-talker = Topic(client, '/chatter', 'std_msgs/String')
+    while client.is_connected:
+        talker.publish(Message({'data': 'Hello World!'}))
+        print('Sending message...')
+        time.sleep(1)
 
-while client.is_connected:
-    talker.publish(Message({'data': 'Hello World!'}))
-    print('Sending message...')
-    time.sleep(1)
-
-talker.unadvertise()
-
-client.terminate()
+    talker.unadvertise()